Move some logic from SpellBuilder into separate SpellList class
