+++ Assets/Scenes/MainScene.unity
@@ -3724,7 +3724,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: c281abe4e37c54142bc7aac9b340b8af, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  audioDict: {fileID: 2077041440}
 --- !u!1 &516854215
 GameObject:
   m_ObjectHideFlags: 0
@@ -3950,6 +3949,45 @@ Transform:
   m_Children: []
   m_Father: {fileID: 1169200567}
   m_RootOrder: 1
+--- !u!1 &537689850
+GameObject:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  serializedVersion: 4
+  m_Component:
+  - 4: {fileID: 537689851}
+  - 114: {fileID: 537689852}
+  m_Layer: 0
+  m_Name: SpellList
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &537689851
+Transform:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 537689850}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: -0.9702144, z: -4.57}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 2017632619}
+  m_RootOrder: 5
+--- !u!114 &537689852
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 537689850}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ce0f5fa29229847b6a0df3e38284c7e4, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &539202623
 GameObject:
   m_ObjectHideFlags: 0
@@ -17378,6 +17416,7 @@ Transform:
   - {fileID: 1806516512}
   - {fileID: 167841923}
   - {fileID: 498005813}
+  - {fileID: 537689851}
   m_Father: {fileID: 0}
   m_RootOrder: 0
 --- !u!1001 &2045182696
