+++ Assets/Scripts/GameEngine/SpellList.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: ce0f5fa29229847b6a0df3e38284c7e4
+timeCreated: 1456286199
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
